fix: unit tests for ExecProgram
